Match OpenAct_800D37F4
Port the title-screen state-machine dispatcher byte-for-byte by
isolating it in its own translation unit and recreating the rdata
layout the original psyq build relied on:

- New file source/overlays/title/onoda/open/openact.c holds OpenAct
  plus the post-OpenAct functions (OpenDie, NewOpen, helpers) and
  the strings that previously lived at 0x800D8848..0x800D9024.
- overlay3.c drops those strings and the reverse-transcribed const-int
  jump tables at 0x800D9028/0x800D9080/0x800D9098 — gcc now re-emits
  equivalent tables at the same addresses.
- linker_command_file.txt links open.obj -> openact.obj -> overlay3.obj
  so openact's rdata fills exactly the gap before the switch tables.
- Splitting OpenAct into its own TU also avoids include_asm_fixup.py
  miscounting an adjacent stub when gcc interleaves switch tables with
  .text in the same TU.
