+++ source/overlays/title/onoda/open/open.c
@@ -1570,148 +1570,3 @@ void title_open_800D3500(OpenWork *work, u_long *ot)
 
     title_open_800C47B8(work, ot);
 }
-
-#pragma INCLUDE_ASM("asm/overlays/title/OpenAct_800D37F4.s")
-void OpenAct_800D37F4(OpenWork *work);
-
-void OpenDie_800D4098(OpenWork *work)
-{
-    int      i;
-    void    *buf;
-
-    GM_FreePrim(work->prim[0]);
-    GM_FreePrim(work->prim[2]);
-    GM_FreePrim(work->prim[3]);
-    GM_FreePrim(work->prim[1]);
-
-    for (i = 0; i < 24; i++)
-    {
-        buf = title_open_800C4B20(&work->kcb[i]);
-        GV_FreeMemory(GV_NORMAL_MEMORY, buf);
-    }
-}
-
-void title_open_800D4174(OpenWork *work, POLY_FT4 *poly, int x0, int y0, int x1, int y1, int abe)
-{
-    setPolyFT4(poly);
-    setRGB0(poly, 128, 128, 128);
-    setXY4(poly, x0, y0, x1, y0, x0, y1, x1, y1);
-    SetSemiTrans(poly, abe);
-}
-
-static inline void title_open_helper_800D41E4(POLY_FT4 *poly, DG_TEX *tex, int uo, int vo)
-{
-    int u0, u1;
-    int v0, v1;
-
-    u0 = tex->off_x;
-    u1 = u0 + tex->w + uo;
-    v0 = tex->off_y;
-    v1 = v0 + tex->h + vo;
-    setUV4(poly, u0, v0, u1, v0, u0, v1, u1, v1);
-
-    poly->tpage = tex->tpage;
-    poly->clut = tex->clut;
-}
-
-void title_open_800D41E4(OpenWork *work, int name, POLY_FT4 *poly, int x0, int y0, int x1, int y1, int abe, int type)
-{
-    DG_TEX *tex;
-
-    title_open_800D4174(work, poly, x0, y0, x1, y1, abe);
-
-    tex = DG_GetTexture(name);
-    if (type == 0)
-    {
-        title_open_helper_800D41E4(poly, tex, 1, 1);
-    }
-    else if (type == 1)
-    {
-        title_open_helper_800D41E4(poly, tex, 1, 0);
-    }
-    else if (type == 2)
-    {
-        title_open_helper_800D41E4(poly, tex, 0, 1);
-    }
-    else if (type == 3)
-    {
-        title_open_helper_800D41E4(poly, tex, 0, 0);
-    }
-}
-
-void title_open_800D4368(OpenWork *work, int name, POLY_FT4 *poly, int x0, int y0, int x1, int y1, int abe)
-{
-    DG_TEX *tex;
-    int     u0, u1;
-    int     v0, v1;
-
-    tex = DG_GetTexture(name);
-
-    setPolyFT4(poly);
-
-    u0 = tex->off_x;
-    u1 = u0 + tex->w + 1;
-    v0 = tex->off_y;
-    v1 = v0 + tex->h + 1;
-    setUV4(poly, u0, v0, u1, v0, u0, v1, u1, v1);
-
-    poly->tpage = tex->tpage;
-    poly->clut = tex->clut;
-
-    setRGB0(poly, 0, 0, 0);
-    setXY4(poly, x0, y0, x1, y0, x0, y1, x1, y1);
-    SetSemiTrans(poly, abe);
-}
-
-void title_open_800D4464(OpenWork *work, int name, POLY_GT4 *poly, int x0, int y0, int x1, int y1, int abe)
-{
-    DG_TEX *tex;
-    int     u0, u1;
-    int     v0, v1;
-
-    tex = DG_GetTexture(name);
-
-    setPolyGT4(poly);
-
-    u0 = tex->off_x;
-    u1 = u0 + tex->w + 1;
-    v0 = tex->off_y;
-    v1 = v0 + tex->h + 1;
-    setUV4(poly, u0, v0, u1, v0, u0, v1, u1, v1);
-
-    poly->tpage = tex->tpage;
-    poly->clut = tex->clut;
-
-    setRGB0(poly, 0, 0, 0);
-    setRGB1(poly, 0, 0, 0);
-    setRGB2(poly, 0, 0, 0);
-    setRGB3(poly, 0, 0, 0);
-    setXY4(poly, x0, y0, x1, y0, x0, y1, x1, y1);
-    SetSemiTrans(poly, abe);
-}
-
-#pragma INCLUDE_ASM("asm/overlays/title/OpenGetResources_800D4584.s")
-int  OpenGetResources_800D4584(OpenWork *work, int);
-
-void *NewOpen(int arg0, int arg1)
-{
-    OpenWork *work;
-
-    GM_GameStatus |= STATE_ALL_OFF;
-
-    work = GV_NewActor(EXEC_LEVEL, sizeof(OpenWork));
-    title_dword_800D92D0 = 0;
-
-    if (work != NULL)
-    {
-        GV_SetNamedActor(&work->actor, OpenAct_800D37F4, OpenDie_800D4098, aOpenC);
-
-        if (OpenGetResources_800D4584(work, arg1) < 0)
-        {
-            GV_DestroyActor(&work->actor);
-            return NULL;
-        }
-    }
-
-    return (void *)work;
-}
